requirements.txt
dj-database-url==0.4.2
Django==1.11
gunicorn==19.7.1
psycopg2==2.7.1
whitenoise==3.3.0
djangorestframework==3.6.2
